requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
yfinance          # For historical prices and stock info
tavily-python     # Optimized search engine for LLM agents
pandas            # Data manipulation

langchain-core
langchain-community
langchain-text-splitters
langchain-classic
langchain-chroma


# --- PDF & Web Loading ---
pymupdf

# --- AI & LLM ---
openai
tiktoken

matplotlib
plotly
langgraph
smolagents
liteLLM
autogen